break out a findCopySites function
Make checkPkg a little smaller by breaking out a findCopySites function.

At the same time, do the types.Func cast in the same loop as the wide
struct calculation and pass the resulting slice of Funcs to
findCopySites.

This isn't much of a performance improvement in a moderately large
codebase on a modern machine, but the additional clarity is nice. If we
have to refactor findCopySites again or add a function for finding wide
structs, it might be worth undoing.